refactor(app): move RobotDashboard
