+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -29,7 +29,7 @@ import { EmergencyStop } from '../pages/EmergencyStop'
 import { NameRobot } from '../pages/NameRobot'
 import { NetworkSetupMenu } from '../pages/NetworkSetupMenu'
 import { ProtocolSetup } from '../pages/ODD/ProtocolSetup'
-import { RobotDashboard } from '../pages/RobotDashboard'
+import { RobotDashboard } from '../pages/ODD/RobotDashboard'
 import { RobotSettingsDashboard } from '../pages/RobotSettingsDashboard'
 import { ProtocolDashboard } from '../pages/ODD/ProtocolDashboard'
 import { ProtocolDetails } from '../pages/ODD/ProtocolDetails'
