+++ app/src/pages/ODD/RobotDashboard/WelcomeModal.tsx
@@ -12,10 +12,10 @@ import {
 } from '@opentrons/components'
 import { useCreateLiveCommandMutation } from '@opentrons/react-api-client'
 
-import { SmallButton } from '../../atoms/buttons'
-import { OddModal } from '../../molecules/OddModal'
+import { SmallButton } from '../../../atoms/buttons'
+import { OddModal } from '../../../molecules/OddModal'
 
-import welcomeModalImage from '../../assets/images/on-device-display/welcome_dashboard_modal.png'
+import welcomeModalImage from '../../../assets/images/on-device-display/welcome_dashboard_modal.png'
 
 import type { SetStatusBarCreateCommand } from '@opentrons/shared-data'
