+++ app/src/pages/ODD/RobotDashboard/__tests__/WelcomeModal.test.tsx
@@ -4,13 +4,13 @@ import { fireEvent, screen } from '@testing-library/react'
 
 import { useCreateLiveCommandMutation } from '@opentrons/react-api-client'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 import { WelcomeModal } from '../WelcomeModal'
 
 import type { SetStatusBarCreateCommand } from '@opentrons/shared-data'
 
-vi.mock('../../../redux/config')
+vi.mock('../../../../redux/config')
 vi.mock('@opentrons/react-api-client')
 
 const mockFunc = vi.fn()
